% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanList}
\alias{cleanList}
\title{Helper function to validate and clean lists}
\usage{
cleanList(lst)
}
\arguments{
\item{lst}{A list that might contain NULL or NA values.}
}
\value{
A list where NULL and NA values have been removed,
or an empty list if the input wasn't a list.
}
\description{
This function takes a list as an argument, checks if it's not NULL and is a list,
and then removes any NA or NULL values from it. If the input is not a list,
it returns an empty list.
}
