% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divide_dimension_value}
\alias{divide_dimension_value}
\title{Divide Dimension Value}
\usage{
divide_dimension_value(value, divisor)
}
\arguments{
\item{value}{A character or numeric value representing the dimension.
It can be in the format "100px", '100\%', or 100.}

\item{divisor}{A numeric value by which the dimension value will be divided.}
}
\value{
A character or numeric value representing the divided dimension.
}
\description{
This function takes a dimension value (which can be in the format '100px', '100\%', or 100)
and divides it by a given divisor. The result is returned in the same format as the input.
}
\examples{
divide_dimension_value('100px', 2)  # Returns "50px"
divide_dimension_value('100\%', 2)  # Returns '50\%'
divide_dimension_value(100, 2)     # Returns 50

}
