% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erythromycin_cluster}
\alias{erythromycin_cluster}
\title{Erythromycin Gene Cluster Data}
\format{
A data frame with 23 rows and 6 columns:
\describe{
  \item{Identifiers}{Unique identifiers for the genes. A character vector.}
  \item{Start}{Start positions of the genes on the chromosome. A numeric vector.}
  \item{End}{End positions of the genes on the chromosome. A numeric vector.}
  \item{Strand}{Strand orientation of the gene, either "+" or "-". A character vector.}
  \item{Product}{Description of the gene products. A character vector.}
  \item{Functions}{Functional categorization of the genes. A character vector.}
}
}
\usage{
erythromycin_cluster
}
\description{
This dataset contains information about various genes involved in the
biosynthesis of Erythromycin, an antibiotic produced by the bacterium
Saccharopolyspora erythraea. It includes details such as gene identifiers,
start and end positions of genes, and their associated functions and products.
}
\keyword{datasets}
