% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_blast.R
\name{GC_blast}
\alias{GC_blast}
\title{Perform Gene Cluster BLAST Analysis}
\usage{
GC_blast(GC_chart, id_column, query, identity = 0.3, ...)
}
\arguments{
\item{GC_chart}{A list object representing the GC chart, which must contain
data and configuration for clusters.}

\item{id_column}{The name of the column that contains the gene identifiers.}

\item{query}{The name of the query cluster to be used for BLAST comparisons.}

\item{identity}{Minimum identity threshold for BLAST hits to be considered
significant. Defaults to 0.3.}

\item{...}{Additional arguments passed to the `protein_blast` function.}
}
\value{
Returns the modified GC_chart object with updated BLAST analysis data.
}
\description{
This function performs a BLAST analysis on gene clusters provided in a GC
chart object. It checks for sequence or translation data, performs protein
BLASTs, and updates the GC chart with BLAST results including identity,
similarity.
}
\note{
This function requires the Biostrings package to be installed. It also
 updates the GC_chart object in place, adding BlastP results to the chart
 configuration.
}
\examples{
#' \dontrun{
GC_chart(
   data,
   cluster = "cluster",
   group = "protein_id",
   height = "900px") \%>\%
   GC_blast(
   query = "V792X672X442M18_1",
   id_column = "protein_id",
   identity = 0.3
   ) \%>\%
   GC_labels() \%>\%
   GC_links("BlastP")
}

}
