% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_cluster}
\alias{GC_cluster}
\title{Modify Gene Track}
\usage{
GC_cluster(
  GC_chart,
  prevent_gene_overlap = FALSE,
  overlap_spacing = 40,
  cluster = NULL,
  style = list(),
  ...
)
}
\arguments{
\item{GC_chart}{The gene chart object to be modified.}

\item{prevent_gene_overlap}{Logical, whether to include the gene track or not.
If FALSE, the specified gene track is removed.}

\item{overlap_spacing}{Numeric, the spacing to be used between overlapping genes.}

\item{cluster}{Numeric or character, specifies the cluster to filter genes by.}

\item{style}{A list of CSS styles to be applied to the gene track.
Each element of the list should be a valid CSS property-value
pair. For example, list(backgroundColor = "red", color = "white").}

\item{...}{Additional arguments to be passed to the underlying functions.}
}
\value{
Returns the modified gene chart object.
}
\description{
This function can switch prevention of gene overlap on, adjust the spacing between tracks and
alter the styling of specified clusters.
}
\examples{
genes_data <- data.frame(
  start = c(10, 20, 30, 170, 210),
  end = c(200, 150, 180, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5'),
  group = c('A', 'A', 'A', 'A', 'A')
)


GC_chart(genes_data, group = "group", height = "150px") \%>\%
  GC_cluster(spacing=30) \%>\%
  GC_legend(FALSE)

}
