% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_item}
\alias{GC_item}
\title{Update Settings of a GC Chart Item}
\usage{
GC_item(GC_chart, setting = NULL, cluster = NULL, ...)
}
\arguments{
\item{GC_chart}{A GC chart object to be updated.}

\item{setting}{Character. The setting within the cluster to be modified.}

\item{cluster}{Numeric or character vector. Cluster(s) within the GC chart to be updated.}

\item{...}{Additional arguments to modify the settings of the cluster.}
}
\value{
An updated GC chart object.
}
\description{
This function modifies the settings of specified clusters within a GC chart.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)

# Update style of a specific cluster
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
GC_labels("name") \%>\%
GC_item(
"labels",
cluster = 2,  # If NULL all clusters will be updated
fontSize = "16px"
)

# Update style of specific items within a specific cluster
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
GC_labels("name") \%>\%
GC_item(
"labels",
cluster = 2, # Can be index or name of cluster
itemStyle =
 list(
   list(index = 0, fontSize = "10px"), # Note index 0 is used
   list(index = 2, fontSize = "16px"). # because of JS array.
   )
 )

}
