% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_sequence}
\alias{GC_sequence}
\title{Update Sequence Display of a GC Chart Cluster}
\usage{
GC_sequence(GC_chart, show = TRUE, cluster = NULL, marker = list(), ...)
}
\arguments{
\item{GC_chart}{A GC chart object.}

\item{show}{Logical. Whether to display the sequence. Default is TRUE.}

\item{cluster}{Numeric or character vector. Clusters in the GC chart to update.}

\item{marker}{List. Settings for the sequence break marker.}

\item{...}{Additional customization arguments for sequence display.}
}
\value{
Updated GC chart with new sequence display settings.
}
\description{
Modify the sequence display of specified clusters within a GC chart and adjust
the display settings.
}
\examples{
genes_data <- data.frame(
  start = c(10, 90, 130, 170, 210),
  stop = c(40, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 2, 2, 2)
)

# Basic usage
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
GC_labels("name") \%>\%
GC_sequence(
  show = TRUE,
  cluster = 1
)

# Customize sequence style
GC_chart(genes_data, cluster="cluster", group = "group") \%>\%
GC_sequence(
  stroke = "black",
  strokeWidth = 2,
  # Any other CSS styles
  marker =         # Style sequence
  list(            # break marker
    stroke = "black",
    strokeWidth = 2,
    gap = 3,
    tiltAmount = 5
    )
  )

}
