% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_title}
\alias{GC_title}
\title{Update Title of a GC Chart Cluster}
\usage{
GC_title(
  GC_chart,
  title = NULL,
  subtitle = NULL,
  subtitleFont = list(),
  titleFont = list(),
  show = TRUE,
  height = "40px",
  cluster = NULL,
  ...
)
}
\arguments{
\item{GC_chart}{A GC chart object.}

\item{title}{Character vector. Titles to set for the clusters.}

\item{subtitle}{Character vector. Subtitles to set for the clusters.}

\item{subtitleFont}{List. Settings for the subtitle font.}

\item{titleFont}{List. Settings for the title font.}

\item{show}{Logical. Whether to display the title. Default is TRUE.}

\item{height}{Character. Height for the title (e.g., "40px").}

\item{cluster}{Numeric or character vector. Clusters in the GC chart to update.}

\item{...}{Additional customization arguments for title and subtitle.}

\item{x}{Numeric. X-coordinate for title's position.}

\item{y}{Numeric. Y-coordinate for title's position.}

\item{position}{Character. Title position ("center", "left", or "right").}

\item{spacing}{Numeric. Spacing between title and subtitle.}
}
\value{
Updated GC chart with new title settings.
}
\description{
Modify the title and subtitle of specified clusters within a GC chart and adjust
the display settings.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)

# Basic usage
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
GC_labels("name") \%>\%
GC_title(
  title = "Cluster 1 Data",
  subtitle = "Detailed View",
  show = TRUE,
  cluster = 1
)

# Customizing title style
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
  GC_labels("name") \%>\%
  GC_title(
    title = "Cluster 1 Data",
    subtitle = "Detailed View",
    show = TRUE,
    cluster = 1,
    x = 0,
    y = 0,
    position = "center",
    spacing = 20,
    titleFont = list(
      size = "16px",
      style = "normal",
      weight = "bold",
      decoration = "normal",
      family = "sans-serif",
      cursor = "default",
      fill = "black"
    ),
    subtitleFont = list(
      size = "14px",
      style = "normal",
      weight = "bold",
      decoration = "normal",
      family = "sans-serif",
      cursor = "default",
      fill = "black"
    )
  )

}
