% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_track}
\alias{GC_track}
\title{Modify Gene Track}
\usage{
GC_track(GC_chart, track = TRUE, spacing = 40, cluster = NULL)
}
\arguments{
\item{GC_chart}{The gene chart object to be modified.}

\item{track}{Logical, whether to include the gene track or not. If FALSE, the specified gene track is removed.}

\item{spacing}{Numeric, the spacing to be used between gene tracks.}

\item{cluster}{Numeric or character, the specific cluster to filter genes by.}
}
\value{
Returns the modified gene chart object.
}
\description{
This function can switch gene tracks off or adjust the spacing between tracks.
}
\examples{
genes_data <- data.frame(
  start = c(10, 20, 30, 170, 210),
  stop = c(200, 150, 180, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5'),
  group = c('A', 'A', 'A', 'A', 'A')
)


GC_chart(genes_data, group = "group", height = "150px") \%>\%
  GC_track(spacing=30) \%>\%
  GC_legend(FALSE)

}
