% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{get_links}
\alias{get_links}
\title{Get Links from GC Chart}
\usage{
get_links(GC_chart, group, value1 = NULL, value2 = NULL, cluster = NULL)
}
\arguments{
\item{GC_chart}{Gene chart object.}

\item{group}{The name of the column in the data to create value pairs from.}

\item{value1}{Optional vector of group values to generate links for.}

\item{value2}{Optional vector of group values to generate links for.}

\item{cluster}{Numeric or character vector or NULL; specifies which clusters
to generate links between.}
}
\value{
A data frame of links between clusters based on the group column.
}
\description{
Processes a `GC_chart` object to create links between clusters. It creates
links for all values in the group column or between value1 and value2 when
specified.
}
\examples{
# See examples for the `GC_links` function for usage.
}
\seealso{
* [GC_links()]
}
