% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scale_breaks}
\alias{get_scale_breaks}
\title{Identify Non-Coding Regions Between Genes}
\usage{
get_scale_breaks(genes, threshold_percentage = 0, padding = 0)
}
\arguments{
\item{genes}{A data.frame containing the start and end positions of genes.
The data.frame must contain columns named 'start' and 'stop'.}

\item{threshold_percentage}{Numeric. A threshold value given as percentage. Only non-coding
regions that are larger than this threshold (relative to the entire genomic region covered)
will be returned.}

\item{padding}{Numeric. A value given as percentage that will be added to the start
and subtracted from the end of each non-coding region.}
}
\value{
A data.frame containing the start and stop positions of the identified non-coding regions.
}
\description{
Given a set of gene positions, this function calculates and returns the non-coding
regions, potentially with added padding. It also considers overlapping genes and ensures
that the starting position is always less than the ending position for each gene.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90),
  stop = c(40, 80, 120)
)
get_scale_breaks(genes_data, threshold_percentage = 5, padding = 2)

}
