% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUpdatedClusters}
\alias{getUpdatedClusters}
\title{Get clusters to update from GCVieweR Object}
\usage{
getUpdatedClusters(GCVieweR, cluster)
}
\arguments{
\item{GCVieweR}{An object containing the series of clusters.}

\item{cluster}{A numeric vector or character vector specifying the clusters to be retrieved.
If NULL (default), all clusters are returned.}
}
\value{
A character vector of cluster names.
}
\description{
This function retrieves the clusters to be updated from a GCVieweR object based on the provided cluster argument.
The function checks if the cluster argument is valid and returns the corresponding clusters.
}
