% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_hox_genes}
\alias{human_hox_genes}
\title{Human HOX Gene Cluster Data}
\format{
A data frame with 39 rows and 5 columns:
\describe{
  \item{symbol}{HGNC symbols for the HOX genes. A character vector.}
  \item{chromosome_name}{Chromosome number where each gene is located. A numeric vector.}
  \item{start}{Start positions of the genes on the chromosome. A numeric vector.}
  \item{end}{End positions of the genes on the chromosome. A numeric vector.}
  \item{cluster}{Cluster identification as a character vector,
  specifying the HOX gene cluster (HOXA, HOXB, HOXC, HOXD).}
  \item{name}{Simplified names for the HOX genes. A character vector.}
}
}
\usage{
human_hox_genes
}
\description{
This dataset represents the positions and clusters of various human HOX
genes. HOX genes are a group of related genes that control the body plan of
an embryo along the head-tail axis. The dataset includes genes from HOXA,
HOXB, HOXC, and HOXD clusters.
}
\keyword{datasets}
