% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ophA_clusters}
\alias{ophA_clusters}
\title{ophA Gene Cluster from Omphalotus olearius}
\format{
A data frame with 17 rows and 5 columns:
\describe{
  \item{name}{Gene names. A character vector.}
  \item{start}{Start positions of the genes. An integer vector.}
  \item{end}{End positions of the genes. An integer vector.}
  \item{class}{Classifications of the genes. A character vector indicating
  the type of protein or function associated with each gene.}
  \item{cluster}{Cluster identification as a character vector, specifying whether
  the gene belongs to the ophA or dbophA gene cluster.}
}
}
\usage{
ophA_clusters
}
\description{
This dataset represents the ophA gene cluster from Omphalotus olearius and
Dendrothele bispora, involved in the production of omphalotin, a cyclic
N-methylated peptide.
}
\references{
van der Velden NS et al. Autocatalytic backbone N-methylation in
a family of ribosomal peptide natural products. Nat Chem Biol. 2017
Aug;13(8):833-835. doi: 10.1038/nchembio.2393. Epub 2017 Jun 5. PMID:
28581484.
}
\keyword{datasets}
