% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_column_or_return}
\alias{select_column_or_return}
\title{Extract Data from Symbol or Return Vector}
\usage{
select_column_or_return(data, x)
}
\arguments{
\item{data}{A data frame from which a column might be extracted.}

\item{x}{An expression that could be a symbol representing a column name or a call that evaluates to a vector.}
}
\value{
A vector extracted from the data frame based on the column name represented by the symbol,
or a vector if the expression evaluates to one.
}
\description{
This function evaluates the provided expression. If the expression is a symbol,
it treats the symbol as a column name and extracts the corresponding column from the provided data frame.
If the expression is a call that evaluates to a vector (e.g., `as.character(df$column)`),
it returns the evaluated vector.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10)
select_column_or_return(df, a)
select_column_or_return(df, as.character(df$class))

}
