% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_annotation}
\alias{GC_annotation}
\title{Add Annotations to a GC_chart}
\usage{
GC_annotation(
  GC_chart,
  cluster = NULL,
  type = "text",
  x = 0,
  y = 0,
  style = list(),
  ...
)
}
\arguments{
\item{GC_chart}{A GC chart object to which the annotations will be added.}

\item{cluster}{Numeric or character vector specifying the clusters to which annotations should be added.}

\item{type}{Character vector specifying the type of annotations to add. The default is "text".}

\item{x}{Numeric vector specifying the x-coordinate for the annotation's position.}

\item{y}{Numeric vector specifying the y-coordinate for the annotation's position.}

\item{style}{A list of CSS styles to be applied to the annotation.
Each element of the list should be a valid CSS property-value
pair. For example, list(fill = "black", fontSize = "12px").}

\item{...}{Additional parameters for customization of annotations, depending on the type.}
}
\value{
Updated GC chart object with added annotations.
}
\description{
This function adds annotations to specified clusters within a GC chart.
Annotations can be of various types (e.g., text) and are positioned based on provided coordinates.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  end = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)

# Adding a text annotation to a GC chart
GC_chart(genes_data, cluster = "cluster", group = "group", height = "400px") \%>\%
  GC_labels("name") \%>\%
  GC_annotation(
    cluster = 1,
    type = "text",
    x = 100,
    y = 100,
    label = "Example Annotation",
    # Additional customization parameters...
  )

}
