% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_chart}
\alias{GC_chart}
\title{Create a GC Chart Visualization}
\usage{
GC_chart(
  data,
  start_col = "start",
  stop_col = "stop",
  cluster = NULL,
  group = NULL,
  width = "100\%",
  height = "400px",
  elementId = NULL,
  scale_breaks = TRUE,
  scale_break_threshold = 20,
  scale_break_padding = 1
)
}
\arguments{
\item{data}{Data frame with genomic data.}

\item{start_col}{Start positions column. Default: "start".}

\item{stop_col}{Stop positions column. Default: "stop".}

\item{cluster}{Optional cluster column.}

\item{group}{Column to determine gene groups. Used for Aesthetic mapping of genes to colors.}

\item{width}{Chart width, e.g., "100%" or 500.}

\item{height}{Chart height, e.g., "400px" or 300.}

\item{elementId}{Optional widget ID.}

\item{scale_breaks}{Use scale breaks? Default: TRUE.}

\item{scale_break_threshold}{% of full range to determine inter-gene regions for scale breaks.}

\item{scale_break_padding}{% of full range as padding on both sides of a scale break.}
}
\value{
GC chart widget.
}
\description{
Generates an interactive GC chart for genomic data.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)
GC_chart(genes_data, group = "group", cluster = "cluster") \%>\% GC_labels("name")

}
