% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_chart}
\alias{GC_chart}
\title{Create a GC Chart Visualization}
\usage{
GC_chart(
  data,
  start = "start",
  stop = "stop",
  cluster = NULL,
  group = NULL,
  width = "100\%",
  height = "400px",
  background_color = "#0000",
  elementId = NULL,
  scale_breaks = TRUE,
  scale_break_threshold = 20,
  scale_break_padding = 1
)
}
\arguments{
\item{data}{Data frame containing genomic information.}

\item{start}{Column name that indicates start positions. Default is "start".}

\item{stop}{Column name that indicates stop positions. Default is "stop".}

\item{cluster}{Optional column name used for clustering purposes. Default is NULL.}

\item{group}{Column name used for gene grouping to influence color aesthetics.}

\item{width}{Width specification for the chart, such as '100\%' or 500. Default is unspecified.}

\item{height}{Height specification for the chart, such as '400px' or 300. Default is unspecified.}

\item{elementId}{Optional identifier string for the widget. Default is NULL.}

\item{scale_breaks}{Logical flag indicating if scale breaks should be employed. Default is TRUE.}

\item{scale_break_threshold}{Numeric value indicating the threshold percentage of the entire range for determining inter-gene regions suitable for scale breaks. Default is unspecified.}

\item{scale_break_padding}{Numeric value indicating the padding percentage of the entire range on either side of a scale break. Default is unspecified.}
}
\value{
A GC chart widget.
}
\description{
Generates an interactive GC chart for genomic data.
}
\examples{
genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)
GC_chart(genes_data, group = "group", cluster = "cluster") \%>\% GC_labels("name")

}
