% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_links}
\alias{GC_links}
\title{Add Links to GC Chart}
\usage{
GC_links(
  GC_chart,
  links,
  curve = TRUE,
  normal_color = "#969696",
  inverted_color = "#d62728",
  style = list(),
  ...
)
}
\arguments{
\item{GC_chart}{Gene chart object.}

\item{links}{Data frame of links generated by `get_links()` to be added to
the chart.}

\item{curve}{Logical; if `TRUE`, links are curved, otherwise straight.}

\item{normal_color}{Color for the links in their normal state.}

\item{inverted_color}{Color for inverted links.}

\item{style}{List of CSS styles to apply to the links.}

\item{...}{Additional arguments passed to the links.}
}
\value{
Modified `GC_chart` object with added links.
# Add links between all groups in each cluster
chart <- GC_chart(genes_data,
                  cluster = "cluster",
                  height = "200px") %>%
                  GC_labels(label = "group")
links <- get_links(chart, group = "group", value1 = "A", value2 = "B")
chart %>% GC_links(links)

# Add links between group A of cluster 1 and A and B of cluster 2
chart <- GC_chart(genes_data,
                  cluster = "cluster",
                  height = "200px") %>%
                  GC_labels(label = "group")
links <- get_links(chart,
                   group = "group",
                   value1 = c("A", "A"),
                   value2 = c("B", "A"),
                   cluster = c(1,2))
chart %>% GC_links(links)

# Style links
chart <- GC_chart(genes_data,
                  cluster = "cluster",
                  height = "200px"
                  ) %>%
                  GC_labels(label = "group")
links <- get_links(chart, group = "group")
chart %>%
  GC_links(
    links,
    curve = FALSE,
    normal_color = "#1f77b4",
    inverted_color = "#d62728",
    style = list(
      stroke = "black",
     strokeWidth = 0.5,
      fillOpacity = 0.4
      # Any other CSS style
    )
  )
}
\description{
Add links generated by  `get_links` to a `GC_chart` object. Links are added
to the graph by their respective rowIDs.
}
\seealso{
* [get_links()]
}
