% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCVieweR.R
\name{GC_scale}
\alias{GC_scale}
\title{Update Scale of a GC Chart Cluster}
\usage{
GC_scale(
  GC_chart,
  cluster = NULL,
  start = NULL,
  stop = NULL,
  breaks = list(),
  ...
)
}
\arguments{
\item{GC_chart}{A GC chart object.}

\item{cluster}{Numeric or character vector. Clusters in the GC chart to update.}

\item{start}{Numeric vector. Starting points for the scales.}

\item{stop}{Numeric vector. Stopping points for the scales.}

\item{breaks}{List. Settings for the scale breaks.}

\item{...}{Additional arguments for scale settings.}
}
\value{
Updated GC chart with new scale settings.
}
\description{
Modify the scale settings for specified clusters within a GC chart.
}
\examples{
genes_data <- data.frame(
start = c(100, 1000, 2000),
stop = c(150, 1500, 2500),
name = c( 'Gene 4', 'Gene 5', 'Gene 6'),
group = c( 'B', 'A', 'C'),
cluster = c(2, 2, 2)
)

# Set manual start and stop values for scales domain
GC_chart(genes_data, cluster ="cluster", group = "group") \%>\%
GC_scale(
  start = 1,
  stop = 2600,
  breaks =
  list(
    list(start = 160, stop = 900),
    list(start = 1600, stop = 1900)
      )
  )

}
