% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneviewer.R
\name{GC_transcript}
\alias{GC_transcript}
\title{Modify Transcript Characteristics within a Chart}
\usage{
GC_transcript(
  GC_chart,
  group = NULL,
  show = TRUE,
  colorScheme = NULL,
  customColors = NULL,
  transcript = NULL,
  styleExons = list(),
  styleIntrons = list(),
  styleUTRs = list(),
  itemStyleExons = list(),
  itemStyleIntrons = list(),
  itemStyleUTRs = list(),
  labelOptions = list(),
  ...
)
}
\arguments{
\item{GC_chart}{The chart object to be modified.}

\item{group}{Column name used for transcript grouping to influence color
aesthetics.}

\item{show}{Logical, whether to show the transcripts or not.}

\item{colorScheme}{Character or NULL, the name of the color scheme to use.}

\item{customColors}{List or NULL, custom colors to apply to the genes.}

\item{transcript}{Numeric or character, the specific transcript Id to filter
transcripts by.}

\item{styleExons}{List, styles to apply to exons in the chart.}

\item{styleIntrons}{List, styles to apply to introns in the chart.}

\item{styleUTRs}{List, styles to apply to UTRs in the chart.}

\item{itemStyleExons}{List, a list of styles to apply to individual exons in
thechart.}

\item{itemStyleIntrons}{List, a list of styles to apply to individual introns
in the chart.}

\item{itemStyleUTRs}{List, a list of styles to apply to individual UTRs in the
chart.}

\item{labelOptions}{List, options for styling labels such as font size, color,
and position.}

\item{...}{Additional arguments to be passed to the gene options.}
}
\value{
Returns the modified gene chart object.
}
\description{
This function updates the chart with specific characteristics for transcripts
based on the given parameters. It can show/hide transcripts, apply a color
scheme, assign custom colors, filter by cluster, and accept additional
options.
}
\examples{
transcript_data <- data.frame(
  transcript = c("transcript1", "transcript1", "transcript1", "transcript1",
                 "transcript1", "transcript1", "transcript1",
                 "transcript2", "transcript2", "transcript2", "transcript2",
                 "transcript2"),
  type = c("5_utr", "intron", "exon", "intron", "exon", "intron", "3_utr",
           "5_utr", "intron", "exon", "intron", "3_utr"),
  start = c(1, 51, 101, 151, 201, 251, 301,
            1, 51, 101, 151, 301),
  end = c(50, 100, 150, 200, 250, 300, 350,
          50, 100, 150, 300, 350),
  strand = rep("forward", 12)
)

GC_chart(
  transcript_data,
  transcript = "transcript",
  type = "type",
  height = "200px"
) \%>\%
  GC_transcript(
    group = NULL,
    show = TRUE,
    transcript = NULL,
    colorScheme = NULL,
    customColors = NULL,
    styleExons = list(
      show = TRUE,
      strokeWidth = 0,
      cursor = "default",
      marker = "box",
      markerSize = "medium",
      arrowheadWidth = NULL,
      arrowheadHeight = NULL,
      markerHeight = NULL,
      cornerRadius = NULL
      # Any other CSS style
    ),
    styleIntrons = list(
      show = TRUE,
      strokeWidth = 1,
      fill = "none",
      cursor = "default",
      marker = "intron",
      markerSize = "medium",
      arrowheadWidth = NULL,
      arrowheadHeight = NULL,
      markerHeight = NULL,
      cornerRadius = NULL
      # Any other CSS style
    ),
    styleUTRs = list(
      show = TRUE,
      fontSize = "10px",
      fontStyle = "normal",
      fontFamily = "sans-serif",
      cursor = "default",
      color = "black",
      fill = "#FFF",
      strokeWidth = 1,
      marker = "box",
      markerSize = "medium",
      arrowheadWidth = NULL,
      arrowheadHeight = NULL,
      markerHeight = NULL,
      cornerRadius = NULL
      # Any other CSS style
    ),
    labelOptions = list(
      show = TRUE,
      xOffset = 2,
      yOffset = 0,
      fontSize = "12px",
      fontStyle = "normal",
      fontWeight = "normal",
      fontFamily = "sans-serif",
      cursor = "default",
      color = "black"
    ),
    itemStyleExons = list(),
    itemStyleIntrons = list(),
    itemStyleUTRs = list()
  )
# Change the appearance of a specific intron
GC_chart(transcript_data,
         transcript = "transcript",
         type = "type",
         height = "200px"
         ) \%>\%
  GC_transcript(
    transcript = 2,
    itemStyleExons = list(list(index = 0, fill = "red")
    )
  )
}
