% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_dystrophin_transcripts}
\alias{hs_dystrophin_transcripts}
\title{Human Dystrophin Transcripts Data}
\format{
A data frame with 202 rows and 5 columns:
\describe{
  \item{transcript}{Transcript names. A character vector representing
  the names of the transcripts.}
  \item{type}{Transcript type.}
  \item{start}{Start positions. An integer vector showing the starting
  positions of each Exon.}
  \item{end}{End positions. An integer vector showing the ending
  positions of each Exon.}
  \item{length}{Transcript lengths. An integer vector
  indicating the length of each transcript.}
}
}
\usage{
hs_dystrophin_transcripts
}
\description{
This dataset contains the Exon positions on human Dystrophin transcripts
'Dp427p2', 'Dp260-2', 'Dp140', 'Dp116', and 'Dp71'.
}
\keyword{datasets}
